requirements.txt
wandb
ruamel.yaml
tensorly
tensorly-torch
torch-harmonics
matplotlib
opt-einsum
h5py
zarr
torch-scatter